fix: safari freeze
